+++ tests/test_beacon.py
@@ -1,6 +1,7 @@
 """Tests for `beacon` module."""
 from asyncio import AbstractEventLoop
 from typing import Any, Dict, Set
+from uuid import uuid4
 
 import pytest
 from pytest_mock import MockerFixture
@@ -46,8 +47,16 @@ def test_beacon(
     expected_services: Set[str],
 ) -> None:
     """Test beacon."""
-    beacon = Beacon(**beacon_params)
+    # Ugly hack to prevent collisions during parallel tests
+    uuid = uuid4()
+    beacon_params["aliases"] = [
+        a.replace("example", f"example-{uuid}") for a in beacon_params.get("aliases", [])
+    ]  # TODO: Fix me
+    expected_services = {
+        s.replace("example", f"example-{uuid}") for s in expected_services
+    }  # TODO: Fix me
 
+    beacon = Beacon(**beacon_params)
     with raise_keyboard_interrupt(timeout=2):
         beacon.run_forever()