+++ CHANGELOG.md
@@ -7,9 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2021-01-22
+### Changed
+- Blink cli layout.
+- Listen cli layout.
+
 ## [0.1.0] - 2021-01-18
 ### Added
 - First release on PyPI.
 
-[Unreleased]: https://github.com/fedejaure/mdns-beacon/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/fedejaure/mdns-beacon/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/fedejaure/mdns-beacon/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/fedejaure/mdns-beacon/compare/releases/tag/v0.1.0